Refactored login route to split up components into a more modularised form
